Proper handling of vector rendering settings

+++ src/mame.h
@@ -199,7 +199,7 @@ struct GameOptions
         
 	int		 vector_width;	        /* requested width for vector games; 0 means default (640) */
 	int		 vector_height;	        /* requested height for vector games; 0 means default (480) */
-	int		 beam;			        /* vector beam width */
+	float	 beam;			        /* vector beam width */
 	float	 vector_flicker;	    /* vector beam flicker effect control */
 	float	 vector_intensity;      /* vector beam intensity */
 	int		 translucency;	        /* 1 to enable translucency on vectors */
